+++ logger/package.go
@@ -1,2 +1,0 @@
-//Package logger contains utilities to log koazee
-package logger