+++ stream/removeduplicates_test.go
@@ -4,8 +4,6 @@ import (
 	"testing"
 
 	"github.com/wesovilabs/koazee"
-	"github.com/wesovilabs/koazee/errors"
-
 	"github.com/wesovilabs/koazee/stream"
 
 	"github.com/stretchr/testify/assert"
@@ -35,13 +33,10 @@ func TestStream_RemoveDuplicates(t *testing.T) {
 		t,
 		[]*person{{"John", 50}},
 		stream.New([]*person{{"John", 50}, {"John", 50}}).RemoveDuplicates().Out().Val())
-}
 
-func TestStream_RemoveDuplicates_validation(t *testing.T) {
-
-	assert.Equal(
-		t,
-		errors.EmptyStream(stream.OpCodeRemoveDuplicates, "A nil Stream can not be iterated"),
-		koazee.Stream().RemoveDuplicates().Out().Err())
-
+	counter, _ := koazee.Stream().RemoveDuplicates().Count()
+	assert.Equal(t,
+		0,
+		counter,
+	)
 }